Initial working version of BioGenie
