+++ bio-genie/frontend/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>BioGenie: Ask a Biology Question</title>
+  <style>
+    body { font-family: sans-serif; padding: 2rem; max-width: 600px; margin: auto; }
+    input, button { padding: 10px; margin-top: 10px; width: 100%; }
+    #answer { margin-top: 20px; font-weight: bold; }
+  </style>
+</head>
+<body>
+  <h2>🔬 Ask BioGenie!</h2>
+  <input type="text" id="question" placeholder="e.g. What is DNA?">
+  <button onclick="ask()">Ask</button>
+  <div id="answer"></div>
+
+  <script>
+    async function ask() {
+      const question = document.getElementById("question").value;
+      const res = await fetch("http://127.0.0.1:8000/ask", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ question })
+      });
+      const data = await res.json();
+      document.getElementById("answer").innerText = data.answer || "No answer.";
+    }
+  </script>
+</body>
+</html>
